Use the amsterdam-jekyll-theme as remote theme
Fixes #57

In order to easily re-use the styling from the Amsterdam Open Source
project in other projects such as the documentation of multiple newly
developed codebases (ie GOB).

The design is now a standalone theme that lives in it's ows repository:

https://github.com/Amsterdam/amsterdam-jekyll-theme

This change changes nothing in the generated files while adding flexibility to the
development process.